Integrate the swift compiler as a built-in tool
