+++ lib/BuildSystem/CMakeLists.txt
@@ -9,5 +9,4 @@ add_llbuild_library(llbuildBuildSystem
   BuildValue.cpp
   ExternalCommand.cpp
   LaneBasedExecutionQueue.cpp
-  SwiftTools.cpp
   )
